requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
joblib==1.3.2
pyyaml==6.0.2
regex==2024.9.11
langdetect==1.0.9
matplotlib==3.8.4
pyarrow==15.0.2
pydantic==2.7.1
Unidecode>=1.3.8
structlog>=24.1.0
rapidfuzz==3.9.6

